requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
cryptography==42.0.8
customtkinter==5.2.2
cytoolz==0.12.3
darkdetect==0.8.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
hexbytes==0.3.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.13.1
libclang==18.1.1
lru-dict==1.2.0
Markdown==3.6
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parsimonious==0.10.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.3
py-solc-x==2.0.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.0.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
web3==6.20.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
